requirements.txt
discord.py==2.4.0
aiohttp==3.10.9
requests==2.32.3
g4f==0.3.2.9
numpy==2.2.4
opencv-python-headless==4.11.0.86
Pillow==11.2.1
easyocr==1.7.2
curl-cffi==0.13.0
nest_asyncio==1.6.0
